backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
sqlmodel>=0.0.14
litellm>=1.0.0
structlog>=24.1.0
rich>=13.7.0
pyjwt[crypto]>=2.8.0
gunicorn>=21.2.0
pytest>=8.0.0
slowapi
ruff

